refactor: move keyword suggester to utils
